rename augur_operations file to operations
